+++ MuxyMobile/VCSView.swift
@@ -432,466 +432,3 @@ private struct StatusBadge: View {
         }
     }
 }
-
-struct BranchesSheet: View {
-    let projectID: UUID
-    let onChange: () async -> Void
-    @Environment(ConnectionManager.self) private var connection
-    @Environment(\.dismiss) private var dismiss
-
-    @State private var branches: VCSBranchesDTO?
-    @State private var isLoading = false
-    @State private var errorMessage: String?
-    @State private var busyBranch: String?
-    @State private var showingCreate = false
-    @State private var newBranchName = ""
-
-    var body: some View {
-        NavigationStack {
-            ZStack {
-                themeBg.ignoresSafeArea()
-                content
-            }
-            .navigationTitle("Branches")
-            .navigationBarTitleDisplayMode(.inline)
-            .toolbarColorScheme(preferredScheme, for: .navigationBar)
-            .tint(themeFg)
-            .toolbar {
-                ToolbarItem(placement: .topBarLeading) {
-                    Button("Close") { dismiss() }.foregroundStyle(themeFg)
-                }
-                ToolbarItem(placement: .topBarTrailing) {
-                    Button { showingCreate = true } label: {
-                        Image(systemName: "plus").foregroundStyle(themeFg)
-                    }
-                }
-            }
-            .alert("New Branch", isPresented: $showingCreate) {
-                TextField("branch-name", text: $newBranchName)
-                Button("Cancel", role: .cancel) { newBranchName = "" }
-                Button("Create") {
-                    let name = newBranchName
-                    newBranchName = ""
-                    Task { await createBranch(name: name) }
-                }
-            } message: {
-                Text("Creates and switches to a new branch from HEAD.")
-            }
-        }
-        .task { await load() }
-    }
-
-    @ViewBuilder
-    private var content: some View {
-        if let branches {
-            List {
-                ForEach(branches.locals, id: \.self) { branch in
-                    branchRow(branch, current: branches.current)
-                }
-                if let error = errorMessage {
-                    Text(error).font(.footnote).foregroundStyle(.red)
-                        .listRowBackground(themeFg.opacity(0.06))
-                }
-            }
-            .scrollContentBackground(.hidden)
-        } else if isLoading {
-            ProgressView().tint(themeFg)
-        } else {
-            Text(errorMessage ?? "No branches").foregroundStyle(themeFg.opacity(0.7))
-        }
-    }
-
-    private func branchRow(_ branch: String, current: String) -> some View {
-        Button {
-            guard branch != current else { return }
-            Task { await switchTo(branch) }
-        } label: {
-            HStack {
-                Image(systemName: branch == current ? "checkmark.circle.fill" : "circle")
-                    .foregroundStyle(branch == current ? .green : themeFg.opacity(0.4))
-                Text(branch)
-                    .foregroundStyle(themeFg)
-                Spacer()
-                if busyBranch == branch {
-                    ProgressView().tint(themeFg)
-                }
-            }
-        }
-        .listRowBackground(themeFg.opacity(0.06))
-    }
-
-    private func load() async {
-        isLoading = true
-        errorMessage = nil
-        do {
-            branches = try await connection.listBranches(projectID: projectID)
-        } catch {
-            errorMessage = error.localizedDescription
-        }
-        isLoading = false
-    }
-
-    private func switchTo(_ branch: String) async {
-        busyBranch = branch
-        defer { busyBranch = nil }
-        do {
-            try await connection.switchBranch(projectID: projectID, branch: branch)
-            await onChange()
-            dismiss()
-        } catch {
-            errorMessage = error.localizedDescription
-        }
-    }
-
-    private func createBranch(name: String) async {
-        let trimmed = name.trimmingCharacters(in: .whitespaces)
-        guard !trimmed.isEmpty else { return }
-        busyBranch = trimmed
-        defer { busyBranch = nil }
-        do {
-            try await connection.createBranch(projectID: projectID, name: trimmed)
-            await onChange()
-            dismiss()
-        } catch {
-            errorMessage = error.localizedDescription
-        }
-    }
-
-    private var themeFg: Color { connection.deviceTheme?.fgColor ?? .primary }
-    private var themeBg: Color { connection.deviceTheme?.bgColor ?? Color(.systemBackground) }
-    private var preferredScheme: ColorScheme { (connection.deviceTheme?.isDark ?? true) ? .dark : .light }
-}
-
-struct WorktreesSheet: View {
-    let projectID: UUID
-    let onChange: () async -> Void
-    @Environment(ConnectionManager.self) private var connection
-    @Environment(\.dismiss) private var dismiss
-
-    @State private var errorMessage: String?
-    @State private var busyID: UUID?
-    @State private var showingAdd = false
-
-    private var activeWorktreeID: UUID? {
-        connection.workspace?.worktreeID
-    }
-
-    var body: some View {
-        NavigationStack {
-            ZStack {
-                themeBg.ignoresSafeArea()
-                content
-            }
-            .navigationTitle("Worktrees")
-            .navigationBarTitleDisplayMode(.inline)
-            .toolbarColorScheme(preferredScheme, for: .navigationBar)
-            .tint(themeFg)
-            .toolbar {
-                ToolbarItem(placement: .topBarLeading) {
-                    Button("Close") { dismiss() }.foregroundStyle(themeFg)
-                }
-                ToolbarItem(placement: .topBarTrailing) {
-                    Button { showingAdd = true } label: {
-                        Image(systemName: "plus").foregroundStyle(themeFg)
-                    }
-                }
-            }
-            .sheet(isPresented: $showingAdd) {
-                AddWorktreeSheet(projectID: projectID)
-            }
-        }
-    }
-
-    @ViewBuilder
-    private var content: some View {
-        let worktrees = connection.projectWorktrees[projectID] ?? []
-        List {
-            ForEach(worktrees) { worktree in
-                row(worktree)
-            }
-            if let error = errorMessage {
-                Text(error).font(.footnote).foregroundStyle(.red)
-                    .listRowBackground(themeFg.opacity(0.06))
-            }
-        }
-        .scrollContentBackground(.hidden)
-    }
-
-    private func row(_ worktree: WorktreeDTO) -> some View {
-        let isActive = worktree.id == activeWorktreeID
-        return Button {
-            guard !isActive else { return }
-            Task { await switchTo(worktree) }
-        } label: {
-            HStack {
-                Image(systemName: isActive ? "checkmark.circle.fill" : (worktree.isPrimary ? "house" : "square.stack.3d.up"))
-                    .foregroundStyle(isActive ? .green : themeFg.opacity(0.7))
-                VStack(alignment: .leading, spacing: 2) {
-                    Text(worktree.name)
-                        .foregroundStyle(themeFg)
-                    if let branch = worktree.branch {
-                        Text(branch)
-                            .font(.caption)
-                            .foregroundStyle(themeFg.opacity(0.6))
-                    }
-                }
-                Spacer()
-                if busyID == worktree.id {
-                    ProgressView().tint(themeFg)
-                }
-            }
-        }
-        .swipeActions(edge: .trailing, allowsFullSwipe: false) {
-            if !worktree.isPrimary, !isActive {
-                Button(role: .destructive) {
-                    Task { await remove(worktree) }
-                } label: {
-                    Label("Remove", systemImage: "trash")
-                }
-            }
-        }
-        .listRowBackground(themeFg.opacity(0.06))
-    }
-
-    private func switchTo(_ worktree: WorktreeDTO) async {
-        busyID = worktree.id
-        defer { busyID = nil }
-        do {
-            try await connection.selectWorktree(projectID: projectID, worktreeID: worktree.id)
-            await onChange()
-            dismiss()
-        } catch {
-            errorMessage = error.localizedDescription
-        }
-    }
-
-    private func remove(_ worktree: WorktreeDTO) async {
-        busyID = worktree.id
-        defer { busyID = nil }
-        do {
-            try await connection.removeWorktree(projectID: projectID, worktreeID: worktree.id)
-        } catch {
-            errorMessage = error.localizedDescription
-        }
-    }
-
-    private var themeFg: Color { connection.deviceTheme?.fgColor ?? .primary }
-    private var themeBg: Color { connection.deviceTheme?.bgColor ?? Color(.systemBackground) }
-    private var preferredScheme: ColorScheme { (connection.deviceTheme?.isDark ?? true) ? .dark : .light }
-}
-
-struct AddWorktreeSheet: View {
-    let projectID: UUID
-    @Environment(ConnectionManager.self) private var connection
-    @Environment(\.dismiss) private var dismiss
-
-    @State private var name = ""
-    @State private var branchName = ""
-    @State private var useExistingBranch = false
-    @State private var existingBranches: [String] = []
-    @State private var selectedExisting = ""
-    @State private var inProgress = false
-    @State private var errorMessage: String?
-
-    var body: some View {
-        NavigationStack {
-            ZStack {
-                themeBg.ignoresSafeArea()
-                Form {
-                    Section("Worktree") {
-                        TextField("Name", text: $name)
-                            .foregroundStyle(themeFg)
-                    }
-                    .listRowBackground(themeFg.opacity(0.06))
-
-                    Section("Branch") {
-                        Picker("Source", selection: $useExistingBranch) {
-                            Text("New Branch").tag(false)
-                            Text("Existing").tag(true)
-                        }
-                        .pickerStyle(.segmented)
-
-                        if useExistingBranch {
-                            Picker("Branch", selection: $selectedExisting) {
-                                ForEach(existingBranches, id: \.self) { Text($0).tag($0) }
-                            }
-                        } else {
-                            TextField("new-branch-name", text: $branchName)
-                                .foregroundStyle(themeFg)
-                        }
-                    }
-                    .listRowBackground(themeFg.opacity(0.06))
-
-                    if let error = errorMessage {
-                        Text(error).font(.footnote).foregroundStyle(.red)
-                            .listRowBackground(themeFg.opacity(0.06))
-                    }
-                }
-                .scrollContentBackground(.hidden)
-            }
-            .navigationTitle("Add Worktree")
-            .navigationBarTitleDisplayMode(.inline)
-            .toolbarColorScheme(preferredScheme, for: .navigationBar)
-            .tint(themeFg)
-            .toolbar {
-                ToolbarItem(placement: .topBarLeading) {
-                    Button("Cancel") { dismiss() }.foregroundStyle(themeFg)
-                }
-                ToolbarItem(placement: .topBarTrailing) {
-                    if inProgress {
-                        ProgressView().tint(themeFg)
-                    } else {
-                        Button("Add") { Task { await submit() } }
-                            .foregroundStyle(themeFg)
-                            .disabled(!canSubmit)
-                    }
-                }
-            }
-        }
-        .task { await loadBranches() }
-    }
-
-    private var canSubmit: Bool {
-        guard !name.trimmingCharacters(in: .whitespaces).isEmpty else { return false }
-        if useExistingBranch {
-            return !selectedExisting.isEmpty
-        }
-        return !branchName.trimmingCharacters(in: .whitespaces).isEmpty
-    }
-
-    private func loadBranches() async {
-        do {
-            let branches = try await connection.listBranches(projectID: projectID)
-            existingBranches = branches.locals
-            if selectedExisting.isEmpty { selectedExisting = branches.locals.first ?? "" }
-        } catch {
-            errorMessage = error.localizedDescription
-        }
-    }
-
-    private func submit() async {
-        inProgress = true
-        defer { inProgress = false }
-        let branch = useExistingBranch
-            ? selectedExisting
-            : branchName.trimmingCharacters(in: .whitespaces)
-        do {
-            try await connection.addWorktree(
-                projectID: projectID,
-                name: name.trimmingCharacters(in: .whitespaces),
-                branch: branch,
-                createBranch: !useExistingBranch
-            )
-            dismiss()
-        } catch {
-            errorMessage = error.localizedDescription
-        }
-    }
-
-    private var themeFg: Color { connection.deviceTheme?.fgColor ?? .primary }
-    private var themeBg: Color { connection.deviceTheme?.bgColor ?? Color(.systemBackground) }
-    private var preferredScheme: ColorScheme { (connection.deviceTheme?.isDark ?? true) ? .dark : .light }
-}
-
-struct CreatePRSheet: View {
-    let projectID: UUID
-    let currentBranch: String
-    let onCreated: () async -> Void
-
-    @Environment(ConnectionManager.self) private var connection
-    @Environment(\.dismiss) private var dismiss
-    @Environment(\.openURL) private var openURL
-
-    @State private var title = ""
-    @State private var prBody = ""
-    @State private var baseBranch: String
-    @State private var draft = false
-    @State private var inProgress = false
-    @State private var errorMessage: String?
-
-    init(projectID: UUID, defaultBase: String?, currentBranch: String, onCreated: @escaping () async -> Void) {
-        self.projectID = projectID
-        self.currentBranch = currentBranch
-        self.onCreated = onCreated
-        _baseBranch = State(initialValue: defaultBase ?? "")
-    }
-
-    var body: some View {
-        NavigationStack {
-            ZStack {
-                themeBg.ignoresSafeArea()
-                Form {
-                    Section("Branch") {
-                        HStack {
-                            Text("From")
-                            Spacer()
-                            Text(currentBranch).foregroundStyle(themeFg.opacity(0.7))
-                        }
-                        TextField("Base (e.g. main)", text: $baseBranch)
-                            .foregroundStyle(themeFg)
-                    }
-                    .listRowBackground(themeFg.opacity(0.06))
-
-                    Section("Details") {
-                        TextField("Title", text: $title)
-                            .foregroundStyle(themeFg)
-                        TextField("Body", text: $prBody, axis: .vertical)
-                            .lineLimit(4 ... 10)
-                            .foregroundStyle(themeFg)
-                        Toggle("Draft", isOn: $draft)
-                    }
-                    .listRowBackground(themeFg.opacity(0.06))
-
-                    if let error = errorMessage {
-                        Text(error).font(.footnote).foregroundStyle(.red)
-                            .listRowBackground(themeFg.opacity(0.06))
-                    }
-                }
-                .scrollContentBackground(.hidden)
-            }
-            .navigationTitle("New Pull Request")
-            .navigationBarTitleDisplayMode(.inline)
-            .toolbarColorScheme(preferredScheme, for: .navigationBar)
-            .tint(themeFg)
-            .toolbar {
-                ToolbarItem(placement: .topBarLeading) {
-                    Button("Cancel") { dismiss() }.foregroundStyle(themeFg)
-                }
-                ToolbarItem(placement: .topBarTrailing) {
-                    if inProgress {
-                        ProgressView().tint(themeFg)
-                    } else {
-                        Button("Create") { Task { await submit() } }
-                            .foregroundStyle(themeFg)
-                            .disabled(title.trimmingCharacters(in: .whitespaces).isEmpty)
-                    }
-                }
-            }
-        }
-    }
-
-    private func submit() async {
-        inProgress = true
-        defer { inProgress = false }
-        do {
-            let result = try await connection.createPullRequest(
-                projectID: projectID,
-                title: title.trimmingCharacters(in: .whitespacesAndNewlines),
-                body: prBody,
-                baseBranch: baseBranch.isEmpty ? nil : baseBranch,
-                draft: draft
-            )
-            await onCreated()
-            dismiss()
-            if let url = URL(string: result.url) {
-                openURL(url)
-            }
-        } catch {
-            errorMessage = error.localizedDescription
-        }
-    }
-
-    private var themeFg: Color { connection.deviceTheme?.fgColor ?? .primary }
-    private var themeBg: Color { connection.deviceTheme?.bgColor ?? Color(.systemBackground) }
-    private var preferredScheme: ColorScheme { (connection.deviceTheme?.isDark ?? true) ? .dark : .light }
-}
